Move dragdrop file upload component to own file
