+++ src/components/MultiFileUploadForm.jsx
@@ -16,6 +16,7 @@ import GetDeclaration from '../utils/GetDeclaration';
 import handleAuthErrors from '../utils/API/handleAuthErrors';
 import LoadingSpinner from './LoadingSpinner';
 import { scrollToTop } from '../utils/ScrollToElement';
+import DragDropZone from './MultiFileUpload/DragDropZone';
 
 const FILE_STATUS_PENDING = 'Pending';
 const FILE_STATUS_IN_PROGRESS = 'in progress';
@@ -72,13 +73,13 @@ const MultiFileUploadForm = ({
   urlThisPage,
 }) => {
   const errorSummaryRef = useRef(null);
-  const inputRef = useRef(null);
-  const multiple = true;
+  // const inputRef = useRef(null);
+  // const multiple = true;
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const declarationId = searchParams.get(URL_DECLARATIONID_IDENTIFIER);
   const [disableButtons, setDisableButtons] = useState(false);
-  const [dragActive, setDragActive] = useState(false);
+  // const [dragActive, setDragActive] = useState(false);
   const [errors, setErrors] = useState([]);
   const [filesAddedForUpload, setFilesAddedForUpload] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
@@ -128,16 +129,6 @@ const MultiFileUploadForm = ({
     setIsLoadingFilelist(false);
   };
 
-  const handleDrag = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.type === 'dragenter' || e.type === 'dragover') {
-      setDragActive(true);
-    } else if (e.type === 'dragleave') {
-      setDragActive(false);
-    }
-  };
-
   const scrollToFocusErrors = () => {
     scrollToTop();
     errorSummaryRef?.current?.focus();
@@ -181,29 +172,6 @@ const MultiFileUploadForm = ({
 
       setFilesAddedForUpload([...filesAddedForUpload, ...newFilesForUpload]);
     }
-  };
-
-  // triggers when file(s) are dropped into the input zone
-  const handleDrop = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setDragActive(false);
-    if (e.dataTransfer.files && e.dataTransfer.files[0]) {
-      storeFilesForUpload(e.dataTransfer.files);
-    }
-  };
-
-  // triggers when file(s) are selected with click
-  const handleChange = (e) => {
-    e.preventDefault();
-    if (e.target.files && e.target.files[0]) {
-      storeFilesForUpload(e.target.files);
-    }
-  };
-
-  const onChooseFilesButtonClick = () => {
-    // triggers the input and opens the file browser for selecting files
-    inputRef.current.click();
   };
 
   const updateFileStatus = ({
@@ -384,65 +352,14 @@ const MultiFileUploadForm = ({
           {pageHeading ? <h1 className="govuk-heading-xl">{pageHeading}</h1> : <h1 className="govuk-heading-xl">Upload files</h1>}
         </div>
       </div>
-      <div className="govuk-grid-row">
-        <div className="govuk-grid-column-three-quarters">
-          <form
-            id="multiFileUpload"
-            onDragEnter={handleDrag}
-            onSubmit={(e) => e.preventDefault()}
-          >
-            <fieldset className="govuk-fieldset">
-              <div className={maxFilesError ? 'govuk-form-group govuk-form-group--error' : 'govuk-form-group'}>
-                <p id="multiFileUploadForm-error" className="govuk-error-message">
-                  <span className="govuk-visually-hidden">Error:</span> {maxFilesError}
-                </p>
-                <input
-                  ref={inputRef}
-                  type="file"
-                  id="multiFileUploadInput"
-                  data-testid="multiFileUploadInput"
-                  multiple={multiple}
-                  onChange={handleChange}
-                />
-                <label
-                  className={dragActive ? 'file-upload-dropzone drag-active' : 'file-upload-dropzone'}
-                  htmlFor="multiFileUploadInput"
-                >
-                  <div>
-                    <p className="govuk-body">Drag and drop files here or</p>
-                    <button
-                      className="govuk-button--text"
-                      type="button"
-                      onClick={onChooseFilesButtonClick}
-                      disabled={disableButtons}
-                    >
-                      Choose files
-                    </button>
-                  </div>
-                </label>
-                {dragActive
-                  && (
-                    <div
-                      id="dragFileElement"
-                      onDragEnter={handleDrag}
-                      onDragLeave={handleDrag}
-                      onDragOver={handleDrag}
-                      onDrop={handleDrop}
-                    />
-                  )}
-              </div>
-            </fieldset>
-            <button
-              className="govuk-button govuk-button--secondary"
-              type="button"
-              onClick={onUploadFiles}
-              disabled={disableButtons}
-            >
-              Upload files
-            </button>
-          </form>
-        </div>
-      </div>
+
+      <DragDropZone
+        maxFilesError={maxFilesError}
+        onUploadFiles={onUploadFiles}
+        disableButtons={disableButtons}
+        storeFilesForUpload={storeFilesForUpload}
+      />
+
       <div className="govuk-grid-row">
         <div className="govuk-grid-column-three-quarters">
           {supportingDocumentsList.length > 0 && (
